CONFIG: Move util functions
